+++ packages/edit-site/src/components/list/table.js
@@ -102,10 +102,12 @@ export default function Table( { templateType } ) {
 						</td>
 
 						<td className="edit-site-list-table-column" role="cell">
-							<AddedBy
-								templateType={ templateType }
-								template={ template }
-							/>
+							{ template ? (
+								<AddedBy
+									postType={ template.type }
+									postId={ template.id }
+								/>
+							) : null }
 						</td>
 						<td className="edit-site-list-table-column" role="cell">
 							<Actions template={ template } />