+++ packages/data/CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## Unreleased
 
+## 9.0.0 (2023-03-29)
+
+### Breaking Changes
+
+-   The `registry.register` function will no longer register a store if another instance is registered with the same name.
+
+## 8.6.0 (2023-03-15)
+
 ## 8.5.0 (2023-03-01)
 
 ## 8.4.0 (2023-02-15)